Tidy up after rebase
